refactor: id 타입 변경

+++ app/src/main/java/woowacourse/shopping/data/source/remote/api/ProductsApiService.kt
@@ -19,6 +19,6 @@ interface ProductsApiService {
     @GET("/products/{id}")
     fun getProductById(
         @Header("accept") accept: String = "*/*",
-        @Path("id") id: Int,
+        @Path("id") id: Long,
     ): Call<ProductResponse>
 }
